feat(Filters): split comparable operation handler
