+++ src/Core/Types.Filters/Expressions/Comparable/ComparableInOperationHandler.cs
@@ -1,0 +1,50 @@
+using System;
+using System.Linq.Expressions;
+using HotChocolate.Language;
+using HotChocolate.Utilities;
+
+namespace HotChocolate.Types.Filters.Expressions
+{
+    public class ComparableInOperationHandler
+        : IExpressionOperationHandler
+    {
+        public bool TryHandle(
+            FilterOperation operation,
+            IInputType type,
+            IValueNode value,
+            Expression instance,
+            ITypeConversion converter,
+            out Expression expression)
+        {
+            if (operation.Type == typeof(IComparable)
+                && value is ListValueNode li)
+            {
+                MemberExpression property =
+                    Expression.Property(instance, operation.Property);
+                var parsedValue = type.ParseLiteral(value);
+
+                switch (operation.Kind)
+                {
+                    case FilterOperationKind.In:
+                        expression = FilterExpressionBuilder.In(
+                            property,
+                            operation.Property.PropertyType,
+                            parsedValue);
+                        return true;
+
+                    case FilterOperationKind.NotIn:
+                        expression = FilterExpressionBuilder.Not(
+                            FilterExpressionBuilder.In(
+                                property,
+                                operation.Property.PropertyType,
+                                parsedValue)
+                        );
+                        return true;
+                }
+            }
+
+            expression = null;
+            return false;
+        }
+    }
+}
